requirements.txt
flask
flask-cors
chromadb
sentence-transformers
rapidfuzz
python-dotenv
boto3
requests

pyspellchecker
gunicorn
